backend/requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.15.2
django-cors-headers>=4.3.1
gunicorn>=22.0.0
whitenoise>=6.6.0
python-decouple>=3.8
dj-database-url>=2.1.0
psycopg2-binary>=2.9.9
google-generativeai>=0.8.0
python-dotenv>=1.0.0
